% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinkr_datasets.R
\name{splinkr_datasets}
\alias{splinkr_datasets}
\title{Read speciesLink datasets}
\usage{
splinkr_datasets(filter = NULL)
}
\arguments{
\item{filter}{character vector containing one or more terms to filter datasets}
}
\value{
`data.frame()` with the datasets
}
\description{
Returns a list of all or filtered datasets available in the speciesLink for searching.
}
\examples{
\dontrun{
# return all datasets
splinkr_datasets()

# Filtering datasets
# The terms order defines filtering
# The second term will be searched for in the subset of datasets established
by the first term. The third term will be searched for in the (sub)subset of
the second term and so on.
splinkr_datasets(filter = c("universidade", "xiloteca", "Brasil"))
}
}
