% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinkr_images.R
\name{splinkr_images}
\alias{splinkr_images}
\title{Display or download image records}
\usage{
splinkr_images(imagecode = NULL, scientificName = NULL, path)
}
\arguments{
\item{imagecode}{Carachter vector containing image codes}

\item{scientificName}{Character vector containing one or more single or compound names, without authors}

\item{path}{Path to save image (not implemented yet)}
}
\description{
This function allows viewing the records of the speciesLink network with images.
Future plains to download the images.
}
\examples{
# Display image by image codes
splinkr_images(imagecode = "FLOR0037759")

splinkr_images(imagecode = c("FLOR0037759", "UEC190851"))

# Display image by scientific names
splinkr_images(scientificName = c("Spigelia insignis", "Anathallis kleinii"))

}
