% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinkr_records.R
\name{splinkr_records}
\alias{splinkr_records}
\title{Read speciesLink datasets}
\usage{
splinkr_records(
  scientificName = NULL,
  barcode = NULL,
  catalogNumber = NULL,
  basisOfRecord = NULL,
  collectionCode = NULL,
  collector = NULL,
  collectorNumber = NULL,
  yearCollected = NULL,
  identifiedBy = NULL,
  yearIdentified = NULL,
  kingdom = NULL,
  phylum = NULL,
  class = NULL,
  order = NULL,
  family = NULL,
  typus = NULL,
  country = NULL,
  stateProvince = NULL,
  county = NULL,
  locality = NULL,
  redlist = NULL,
  maxrecords = NULL
)
}
\arguments{
\item{scientificName}{Character vector containing one or more single or compound names, without authors}

\item{barcode}{Character vector containing one or more institution barcode}

\item{catalogNumber}{Character vector containing one or more institution catolog number}

\item{basisOfRecord}{PreservedSpecimen, LivingSpecimen, FossilSpecimen, HumanObservation, MachineObservation, MaterialSample}

\item{collectionCode}{Character vector containing collections or subcollections acronym}

\item{collector}{Character vector containing one or more colletor name}

\item{collectorNumber}{Character vector containing collector number}

\item{yearCollected}{Numeric four-digits collect year}

\item{identifiedBy}{Character vector identificator name}

\item{yearIdentified}{Numeric four-digits year of identifications}

\item{kingdom}{Character vector containing kingdon name}

\item{phylum}{Character vector containing phyllum name}

\item{class}{Character vector containing class name}

\item{order}{Character vector containing order name}

\item{family}{Character vector containing family name}

\item{typus}{If "yes" search only in type material variable}

\item{country}{Character vector containing collect country}

\item{stateProvince}{Character vector containing state}

\item{county}{Character vector containing collect county or city}

\item{locality}{Character vector containingcollect locality}

\item{redlist}{If "yes" return only taxa in Portaria MMA 443/2014}

\item{maxrecords}{Character vector to restrict number of records to return}
}
\value{
`data.frame()` with the records
}
\description{
This function returns the search results from speciesLink network records.
}
\examples{
\dontrun{
# Searching taxon
splinkr_records(scientificName = "Dyckia encholirioides var. rubra")

# Searching taxa collected by collector
splinkr_records(scientificName = c("Anathallis kleinii",
                                   "Anathallis microphyta"),
                collector = "Siqueira")

# Searching all types from
# one collection
splinkr_records(collectionCode = "FLOR", typus = "yes")

# two collections
splinkr_records(collectionCode = c("FLOR", "FURB"), typus = "yes")
}
}
