% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splinkr_datasets.R
\name{splinkr_datasets}
\alias{splinkr_datasets}
\title{Read speciesLink datasets}
\usage{
splinkr_datasets(filter = NULL)
}
\arguments{
\item{filter}{Character vector containing one or more terms to filter datasets}
}
\value{
\code{data.frame()} with the datasets
}
\description{
Returns a \code{tibble} with all or filtered datasets available in the speciesLink for searching.
}
\examples{

# return all datasets
splinkr_datasets()

# Filtering datasets. Terms order defines filtering
# The second term will be searched in subset of datasets created
by the first term. The third term will be searched in the (sub)subset of
the second term and so on.
splinkr_datasets(filter = c("universidade", "xiloteca", "Brasil"))

}
